rename: ファイル名を変更→email_setting_page.dart
