+++ lib/router.dart
@@ -6,7 +6,7 @@ import 'package:todo_app/presentation/pages/auth/login/forgot_password_page.dart
 import 'package:todo_app/presentation/pages/auth/login/login_page.dart';
 import 'package:todo_app/presentation/pages/auth/register/register_page.dart';
 import 'package:todo_app/presentation/pages/home/home_page.dart';
-import 'package:todo_app/presentation/pages/home/profile/section_pages/email_setting.dart';
+import 'package:todo_app/presentation/pages/home/profile/section_pages/email_setting_page.dart';
 import 'package:todo_app/presentation/pages/home/profile/section_pages/terms_and_policies_page.dart';
 import 'package:todo_app/presentation/pages/user_session/user_session_page.dart';
